refactor: move check() function to its own module
