+++ Makefile
@@ -12,7 +12,7 @@ SOURCES = teal/debug.tl teal/attributes.tl teal/errors.tl teal/lexer.tl \
 	teal/gen/lua_generator.tl teal/gen/lua_compat.tl teal/variables.tl teal/type_reporter.tl \
 	teal/macroexps.tl teal/metamethods.tl \
 	teal/type_errors.tl teal/environment.tl \
-	teal/check/context.tl teal/check/visitors.tl \
+	teal/check/context.tl teal/check/visitors.tl teal/check/check.tl \
 	teal/check/relations.tl teal/check/special_functions.tl \
 	teal/check/type_checker.tl teal/check/node_checker.tl \
 	teal/check/file_checker.tl teal/check/string_checker.tl \
