+++ teal/check/visitors.lua
@@ -2,7 +2,7 @@ local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 th
 
 
 local context = require("teal.check.context")
-local Context = context.Context
+
 
 local tldebug = require("teal.debug")
 local TL_DEBUG = tldebug.TL_DEBUG
@@ -77,11 +77,8 @@ local flip_binop_to_metamethod = metamethods.flip_binop_to_metamethod
 
 local traversal = require("teal.traversal")
 
-
 local traverse_nodes = traversal.traverse_nodes
 local fields_of = traversal.fields_of
-
-local type_reporter = require("teal.type_reporter")
 
 local type_errors = require("teal.type_errors")
 
@@ -91,19 +88,19 @@ local variables = require("teal.variables")
 
 
 local util = require("teal.util")
-local shallow_copy_table = util.shallow_copy_table
 local sorted_keys = util.sorted_keys
-
-local environment = require("teal.environment")
-
-
-
 
 
 
 local visitors = {}
 
 
+
+
+visitors.visit_node = {}
+visitors.visit_type = {}
+local visit_node = visitors.visit_node
+local visit_type = visitors.visit_type
 
 
 
@@ -758,8 +755,6 @@ local function assert_is_a(ctx, w, t1, t2, ectx, name)
    end
    return ok
 end
-
-local visit_node = {}
 
 visit_node.cbs = {
    ["statements"] = {
@@ -2383,195 +2378,192 @@ local metamethod_is_method = {
    ["__unm"] = true,
 }
 
-local visit_type
-visit_type = {
-   cbs = {
-      ["generic"] = {
-         before = function(self, typ)
-            self:begin_implied_scope()
-            self:add_var(nil, "@generic", typ)
-         end,
-         after = function(self, typ, _children)
-            self:end_implied_scope()
-            return self:fresh_typeargs(typ)
-         end,
-      },
-      ["function"] = {
-         after = function(self, typ, _children)
-            if self.feat_arity == false then
-               typ.min_arity = 0
-            end
-            return typ
-         end,
-      },
-      ["record"] = {
-         before = function(self, typ)
-            self:begin_implied_scope()
-            self:begin_temporary_record_types(typ)
-         end,
-         after = function(self, typ, children)
-            local i = 1
-            if typ.interface_list then
-               for j, _ in ipairs(typ.interface_list) do
-                  local iface = children[i]
-                  if iface.typename == "array" then
+visit_type.cbs = {
+   ["generic"] = {
+      before = function(self, typ)
+         self:begin_implied_scope()
+         self:add_var(nil, "@generic", typ)
+      end,
+      after = function(self, typ, _children)
+         self:end_implied_scope()
+         return self:fresh_typeargs(typ)
+      end,
+   },
+   ["function"] = {
+      after = function(self, typ, _children)
+         if self.feat_arity == false then
+            typ.min_arity = 0
+         end
+         return typ
+      end,
+   },
+   ["record"] = {
+      before = function(self, typ)
+         self:begin_implied_scope()
+         self:begin_temporary_record_types(typ)
+      end,
+      after = function(self, typ, children)
+         local i = 1
+         if typ.interface_list then
+            for j, _ in ipairs(typ.interface_list) do
+               local iface = children[i]
+               if iface.typename == "array" then
+                  typ.interface_list[j] = iface
+               elseif iface.typename == "nominal" then
+                  local ri = self:resolve_nominal(iface)
+                  if ri.typename == "interface" then
                      typ.interface_list[j] = iface
-                  elseif iface.typename == "nominal" then
-                     local ri = self:resolve_nominal(iface)
-                     if ri.typename == "interface" then
-                        typ.interface_list[j] = iface
-                     else
-                        self.errs:add(children[i], "%s is not an interface", children[i])
+                  else
+                     self.errs:add(children[i], "%s is not an interface", children[i])
+                  end
+               end
+               i = i + 1
+            end
+         end
+         if typ.elements then
+            typ.elements = children[i]
+            i = i + 1
+         end
+         local fmacros
+         local g
+         for name, _ in fields_of(typ) do
+            local ftype = children[i]
+            if ftype.typename == "function" then
+               if ftype.macroexp then
+                  fmacros = fmacros or {}
+                  table.insert(fmacros, ftype)
+               end
+
+               if ftype.is_method then
+                  local fargs = ftype.args.tuple
+                  if fargs[1] then
+                     if not g then
+                        g = self:find_var("@generic")
+                     end
+                     ftype.is_method = ensure_is_method_self(typ, fargs[1], g and g.t)
+                     if ftype.is_method then
+                        fargs[1] = a_type(fargs[1], "self", { display_type = typ })
                      end
                   end
-                  i = i + 1
+               end
+            elseif ftype.typename == "typedecl" and ftype.is_alias then
+               self:resolve_typealias(ftype)
+            end
+
+            typ.fields[name] = ftype
+            i = i + 1
+         end
+         for name, _ in fields_of(typ, "meta") do
+            local ftype = children[i]
+            if ftype.typename == "function" then
+               if ftype.macroexp then
+                  fmacros = fmacros or {}
+                  table.insert(fmacros, ftype)
+               end
+               ftype.is_method = metamethod_is_method[name]
+            end
+            typ.meta_fields[name] = ftype
+            i = i + 1
+         end
+
+         if typ.interface_list then
+            self:expand_interfaces(typ)
+
+            if self.collector then
+               for fname, ftype in fields_of(typ) do
+                  self.env.reporter:add_field(typ, fname, ftype)
                end
             end
-            if typ.elements then
-               typ.elements = children[i]
-               i = i + 1
-            end
-            local fmacros
-            local g
-            for name, _ in fields_of(typ) do
-               local ftype = children[i]
-               if ftype.typename == "function" then
-                  if ftype.macroexp then
-                     fmacros = fmacros or {}
-                     table.insert(fmacros, ftype)
-                  end
+         end
 
-                  if ftype.is_method then
-                     local fargs = ftype.args.tuple
-                     if fargs[1] then
-                        if not g then
-                           g = self:find_var("@generic")
-                        end
-                        ftype.is_method = ensure_is_method_self(typ, fargs[1], g and g.t)
-                        if ftype.is_method then
-                           fargs[1] = a_type(fargs[1], "self", { display_type = typ })
-                        end
-                     end
-                  end
-               elseif ftype.typename == "typedecl" and ftype.is_alias then
-                  self:resolve_typealias(ftype)
-               end
+         if fmacros then
+            for _, t in ipairs(fmacros) do
+               local macroexp_type = traverse_nodes(self, t.macroexp, visit_node, visit_type)
 
-               typ.fields[name] = ftype
-               i = i + 1
-            end
-            for name, _ in fields_of(typ, "meta") do
-               local ftype = children[i]
-               if ftype.typename == "function" then
-                  if ftype.macroexp then
-                     fmacros = fmacros or {}
-                     table.insert(fmacros, ftype)
-                  end
-                  ftype.is_method = metamethod_is_method[name]
-               end
-               typ.meta_fields[name] = ftype
-               i = i + 1
-            end
+               macroexps.check_arg_use(self, t.macroexp)
 
-            if typ.interface_list then
-               self:expand_interfaces(typ)
-
-               if self.collector then
-                  for fname, ftype in fields_of(typ) do
-                     self.env.reporter:add_field(typ, fname, ftype)
-                  end
+               if not self:is_a(macroexp_type, t) then
+                  self.errs:add(macroexp_type, "macroexp type does not match declaration")
                end
             end
+         end
 
-            if fmacros then
-               for _, t in ipairs(fmacros) do
-                  local macroexp_type = traverse_nodes(self, t.macroexp, visit_node, visit_type)
+         self:end_temporary_record_types(typ)
+         self:end_implied_scope()
 
-                  macroexps.check_arg_use(self, t.macroexp)
+         return typ
+      end,
+   },
+   ["typearg"] = {
+      after = function(self, typ, _children)
+         local name = typ.typearg
+         local old = self:find_var(name, "check_only")
+         if old then
+            self.errs:redeclaration_warning(typ, name, "type argument", old)
+         end
+         if simple_types[name] then
+            self.errs:add(typ, "cannot use base type name '" .. name .. "' as a type variable")
+         end
 
-                  if not self:is_a(macroexp_type, t) then
-                     self.errs:add(macroexp_type, "macroexp type does not match declaration")
-                  end
-               end
-            end
-
-            self:end_temporary_record_types(typ)
-            self:end_implied_scope()
-
+         self:add_var(nil, name, a_type(typ, "typearg", {
+            typearg = name,
+            constraint = typ.constraint,
+         }))
+         return typ
+      end,
+   },
+   ["typevar"] = {
+      after = function(self, typ, _children)
+         if not self:find_var_type(typ.typevar) then
+            self.errs:add(typ, "undefined type variable " .. typ.typevar)
+         end
+         return typ
+      end,
+   },
+   ["nominal"] = {
+      after = function(self, typ, _children)
+         if typ.found then
             return typ
-         end,
-      },
-      ["typearg"] = {
-         after = function(self, typ, _children)
-            local name = typ.typearg
-            local old = self:find_var(name, "check_only")
-            if old then
-               self.errs:redeclaration_warning(typ, name, "type argument", old)
-            end
-            if simple_types[name] then
-               self.errs:add(typ, "cannot use base type name '" .. name .. "' as a type variable")
-            end
+         end
 
-            self:add_var(nil, name, a_type(typ, "typearg", {
-               typearg = name,
-               constraint = typ.constraint,
-            }))
-            return typ
-         end,
-      },
-      ["typevar"] = {
-         after = function(self, typ, _children)
-            if not self:find_var_type(typ.typevar) then
-               self.errs:add(typ, "undefined type variable " .. typ.typevar)
-            end
-            return typ
-         end,
-      },
-      ["nominal"] = {
-         after = function(self, typ, _children)
-            if typ.found then
-               return typ
-            end
-
-            local t, typearg = self:find_type(typ.names)
-            if t then
-               local def = t.def
-               if t.is_alias then
-                  if def.typename == "generic" then
-                     def = def.t
-                  end
-                  if def.typename == "nominal" then
-                     typ.found = def.found
-                  end
-               elseif def.typename ~= "circular_require" then
-                  typ.found = t
+         local t, typearg = self:find_type(typ.names)
+         if t then
+            local def = t.def
+            if t.is_alias then
+               if def.typename == "generic" then
+                  def = def.t
                end
-            elseif typearg then
+               if def.typename == "nominal" then
+                  typ.found = def.found
+               end
+            elseif def.typename ~= "circular_require" then
+               typ.found = t
+            end
+         elseif typearg then
 
-               typ.names = nil
-               edit_type(typ, typ, "typevar")
-               local tv = typ
-               tv.typevar = typearg.typearg
-               tv.constraint = typearg.constraint
-            else
-               local name = typ.names[1]
-               local scope = self.st[#self.st]
-               scope.pending_nominals = scope.pending_nominals or {}
-               scope.pending_nominals[name] = scope.pending_nominals[name] or {}
-               table.insert(scope.pending_nominals[name], typ)
-            end
-            return typ
-         end,
-      },
-      ["union"] = {
-         after = function(self, typ, _children)
-            local _, err = is_valid_union(typ)
-            if err then
-               return self.errs:invalid_at(typ, err, typ)
-            end
-            return typ
-         end,
-      },
+            typ.names = nil
+            edit_type(typ, typ, "typevar")
+            local tv = typ
+            tv.typevar = typearg.typearg
+            tv.constraint = typearg.constraint
+         else
+            local name = typ.names[1]
+            local scope = self.st[#self.st]
+            scope.pending_nominals = scope.pending_nominals or {}
+            scope.pending_nominals[name] = scope.pending_nominals[name] or {}
+            table.insert(scope.pending_nominals[name], typ)
+         end
+         return typ
+      end,
+   },
+   ["union"] = {
+      after = function(self, typ, _children)
+         local _, err = is_valid_union(typ)
+         if err then
+            return self.errs:invalid_at(typ, err, typ)
+         end
+         return typ
+      end,
    },
 }
 
@@ -2604,136 +2596,5 @@ visit_type.cbs["any"] = default_type_visitor
 visit_type.cbs["unknown"] = default_type_visitor
 visit_type.cbs["invalid"] = default_type_visitor
 visit_type.cbs["none"] = default_type_visitor
-
-
-
-local function internal_compiler_check(fn)
-   return function(s, n, children, t)
-      t = fn and fn(s, n, children, t) or t
-
-      if type(t) ~= "table" then
-         error(((n).kind or (n).typename) .. " did not produce a type")
-      end
-      if type(t.typename) ~= "string" then
-         error(((n).kind or (n).typename) .. " type does not have a typename")
-      end
-
-      return t
-   end
-end
-
-local function store_type_after(fn)
-   return function(self, n, children, t)
-      t = fn and fn(self, n, children, t) or t
-
-      local w = n
-
-      if w.y then
-         self.collector.store_type(w.y, w.x, t)
-      end
-
-      return t
-   end
-end
-
-local function debug_type_after(fn)
-   return function(s, node, children, t)
-      t = fn and fn(s, node, children, t) or t
-
-      node.debug_type = t
-      return t
-   end
-end
-
-local function patch_visitors(my_visit_node,
-   after_node,
-   my_visit_type,
-   after_type)
-
-
-   if my_visit_node == visit_node then
-      my_visit_node = shallow_copy_table(my_visit_node)
-   end
-   my_visit_node.after = after_node(my_visit_node.after)
-   if my_visit_type then
-      if my_visit_type == visit_type then
-         my_visit_type = shallow_copy_table(my_visit_type)
-      end
-      my_visit_type.after = after_type(my_visit_type.after)
-   else
-      my_visit_type = visit_type
-   end
-   return my_visit_node, my_visit_type
-end
-
-
-visitors.check = function(ast, filename, opts, env)
-   filename = filename or "?"
-
-   if not env then
-      local err
-      env, err = environment.new(opts)
-      if err then
-         return nil, err
-      end
-   end
-   opts = opts or env.defaults
-
-   local self = Context.new(env, filename, opts)
-
-   if env.report_types then
-      env.reporter = env.reporter or type_reporter.new()
-      self.collector = env.reporter:get_collector(filename)
-   end
-
-   local visit_node, visit_type = visit_node, visit_type
-   if opts.run_internal_compiler_checks then
-      visit_node, visit_type = patch_visitors(
-      visit_node, internal_compiler_check,
-      visit_type, internal_compiler_check)
-
-   end
-   if self.collector then
-      visit_node, visit_type = patch_visitors(
-      visit_node, store_type_after,
-      visit_type, store_type_after)
-
-   end
-   if TL_DEBUG then
-      visit_node, visit_type = patch_visitors(
-      visit_node, debug_type_after)
-
-   end
-
-   assert(ast.kind == "statements")
-   traverse_nodes(self, ast, visit_node, visit_type)
-
-   local global_scope = self.st[1]
-   variables.close_types(global_scope)
-   self.errs:check_var_usage(global_scope, true)
-
-   errors.clear_redundant_errors(self.errs.errors)
-
-
-   local result = {
-      ast = ast,
-      env = env,
-      type = self.module_type or a_type(ast, "boolean", {}),
-      filename = filename,
-      warnings = self.errs.warnings,
-      type_errors = self.errs.errors,
-      dependencies = self.dependencies,
-      needs_compat = self.needs_compat,
-   }
-
-   env.loaded[filename] = result
-   table.insert(env.loaded_order, filename or "")
-
-   if self.collector then
-      env.reporter:store_result(self.collector, env.globals)
-   end
-
-   return result
-end
 
 return visitors
